requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
rapidfuzz==3.5.2
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
